+++ src/main/java/com/climingo/climingoApi/auth/exception/ExpiredTokenException.java
@@ -1,4 +1,4 @@
-package com.climingo.climingoApi.security.exception;
+package com.climingo.climingoApi.auth.exception;
 
 import io.jsonwebtoken.JwtException;
