+++ src/main/java/com/climingo/climingoApi/member/api/SignUpController.java
@@ -3,8 +3,8 @@ package com.climingo.climingoApi.member.api;
 import com.climingo.climingoApi.member.api.request.SignUpRequest;
 import com.climingo.climingoApi.member.api.response.MemberInfo;
 import com.climingo.climingoApi.member.application.SignUpService;
-import com.climingo.climingoApi.security.api.response.TokenResponse;
-import com.climingo.climingoApi.security.application.TokenService;
+import com.climingo.climingoApi.auth.api.response.TokenResponse;
+import com.climingo.climingoApi.auth.application.TokenService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
